+++ src/test/java/attendance/domain/CrewStatusTest.java
@@ -14,9 +14,9 @@ public class CrewStatusTest {
     @DisplayName("결석 2회 이상일 경우, 경고를 반환한다")
     @Test
     void given_over_two_absence_then_return_warning() {
-        Map<AttendanceType, Integer> attendanceResult = new HashMap<>();
-        attendanceResult.put(LATE, 3);
-        attendanceResult.put(ABSENCE, 1);
+        Map<AttendanceType, Long> attendanceResult = new HashMap<>();
+        attendanceResult.put(LATE, 3L);
+        attendanceResult.put(ABSENCE, 1L);
         CrewStatus crewStatus = CrewStatus.calculate(AttendanceResult.from(attendanceResult));
         assertThat(crewStatus).isEqualTo(WARING);
     }
@@ -24,9 +24,9 @@ public class CrewStatusTest {
     @DisplayName("결석 3회 이상일 경우, 면담을 반환한다")
     @Test
     void given_over_three_absence_then_return_interview() {
-        Map<AttendanceType, Integer> attendanceResult = new HashMap<>();
-        attendanceResult.put(LATE, 3);
-        attendanceResult.put(ABSENCE, 2);
+        Map<AttendanceType, Long> attendanceResult = new HashMap<>();
+        attendanceResult.put(LATE, 3L);
+        attendanceResult.put(ABSENCE, 2L);
         CrewStatus crewStatus = CrewStatus.calculate(AttendanceResult.from(attendanceResult));
         assertThat(crewStatus).isEqualTo(INTERVIEW);
     }
@@ -34,9 +34,9 @@ public class CrewStatusTest {
     @DisplayName("결석 5회 이상일 경우, 제적을 반환한다")
     @Test
     void given_over_fifth_absence_then_return_absence() {
-        Map<AttendanceType, Integer> attendanceResult = new HashMap<>();
-        attendanceResult.put(LATE, 12);
-        attendanceResult.put(ABSENCE, 1);
+        Map<AttendanceType, Long> attendanceResult = new HashMap<>();
+        attendanceResult.put(LATE, 12L);
+        attendanceResult.put(ABSENCE, 1L);
         CrewStatus crewStatus = CrewStatus.calculate(AttendanceResult.from(attendanceResult));
         assertThat(crewStatus).isEqualTo(FIRE);
     }
